Extract common logic from modal component.
